Asked agent to only import upon login
As it was at face value it imported the users upon login and when-
ever the federated users were listed or searched (most likely on
the admin UI context).

Which sorta accelerates the migration process but in reality is not
what we're looking for, also that bulk load will impact on Keycloaks
performance.