Ask for: Centralize cred. mngm on the respective obj
This one (just as the test generation one) was singularly tricky.

The problem wasn't on the prompt, like prior attempts, but instead
the agent kept taking shortcuts on the dependency injection. It
took over 5 or more "feedback prompts" to get to something close to
the end result.

To be specific these were the "sins" the agent kept falling into:

  - Scope creep (one method doing more than one thing)
  - Obj. Circular dependencies
  - Passing more than was needed (ie: whole user model than a string)
  - Silently adding back code that we deemed as out of place