Asked For: RealmResource for legacy /token endpoint
This one required scope clarification since the agent interpreted
the initial prompt as if we wanted to extend the OIDC token end-
point implementation, when in reality we only needed to reformat
and foward the request.

The other fine tunning we required is to not use sub-resources
since a single one with a @Path should suffice.