Iteration 1 done - Package renaming
This was more a vanity change, since I'll need to repackage all
these components to the real identifier with context unaware tools.

In a nutshell, sed would had gone bunkers if I continued with the
org.keycloak package.

Also, when creating SPI you can use any package name, since it is
the resource/META-INF/services/* files that signal the JVM how to
map each implementation. I bet the package suggestion from the
agent was because Keycloak itself implements a handful of SPIs and
it generated the bias on the naming convention.